perf: Improved the performance caused by `tsc` (#67)
`tsc`, offered by null-ls, caused the performance issue when the usage is
combined with `auto-save`. The `auto-save` was previously set to save
whenever the textChanged or exited the insertMode. This triggers the `tsc`
to compile and send diagnostics at the workspace level every time code has
been updated or changed. Increasing the debounce rate and changing
the save method fixed the performance issues.

It is still necessary to improve the `tsc` checking method in the near
future.