feat: Replaced the plugin and updated the colorScheme
Replaced the `nvim-tree` with `telescope_file_browser`, and updated the
colorScheme's setting. The background of the colorScheme is set to
`transparent=true` and color theme is changed to `terafox` from
`nightfox`.

Note: It may be necessary to change the background color from Terminal
emulator as background color from the colorScheme is set to `transparent`.