feat: Updated the settings and added the `cspell` to `null-ls`
Updated the settings for `lspconfig` to lively update the diagnostics.
This may cause performance issues, and the `virtual_text` is disabled to
reduce performance bottleneck.

Added the `cspell` diagnostics and codeactions to `null-ls`.
Additionally, the custom file directory function for the `cspell.json` is added
to integrate the dictionaries across multiple workspaces. This may increase i
the usability; however, the performance may cause issues later.