feat: Added the `ltex` language tool
Installed the new plugin called `ltex-extra` to improve the usability of `ltex`
lsp. Currently, `ltex` has a feature to create custom path for the dictionary.
If the custom path is found, It will use the custom path. If custom path is not
found, the cwd is taken and dictionary will be created within the root level.
This can lead us to manage all dictionaries, and other ltex's related files in a
single place.

Other minor changes:

1. Refactored the cspell in `null-ls`
2. Add the autocmd to reload ltex dictionary files on saving (or BufWritePost)
3. Changed the `ltex` and `null-ls`'s cspell diagnostics severity to HINT
4. Added the `.gitignore`